+++ libraries/chain/protocol/transaction.cpp
@@ -90,7 +90,7 @@ void transaction::set_expiration( fc::time_point_sec expiration_time )
 
 void transaction::set_reference_block( const block_id_type& reference_block )
 {
-   ref_block_num = fc::endian_reverse_u32(reference_block._hash[0]);
+   ref_block_num = static_cast<uint16_t>(fc::endian_reverse_u32(reference_block._hash[0]));
    ref_block_prefix = reference_block._hash[1];
 }
